+++ tests/nox/test-output/.gitignore
@@ -1,0 +1,1 @@
+*
